fix(number-utils): 변경 한 타입에 맞게 유틸에서 다루는 타입도 높은 숫자를 다루도록 변경

+++ src/main/java/util/NumberUtils.java
@@ -8,11 +8,11 @@ public final class NumberUtils {
     }
 
     public static boolean isZero(final Number number) {
-        return number.intValue() == ZERO;
+        return number.longValue() == ZERO;
     }
 
     public static boolean isNegative(final Number number) {
-        return number.intValue() < ZERO;
+        return number.longValue() < ZERO;
     }
 
     public static int parseInt(final String text) {
